Improve the errors for structuring some of the primitives.
I think that most of the errors right now will be coming from
`structure` function rather than the `unstructure` function. This commit
starts to add some tests for some of the primitives for when they are
given values that they can't deal with.

Now that the primitives are raising more readable errors we can focus on
making the collections have better errors.