Create some docs for the project.
This commit introduces some docs for the project. It does this uses the
tool Sphinx. Sphinx is commonly used within python because it is able to
generate documentation from python docstrings.

This commit adds some very basic docs, updates the docstrings of the
`structure` and `unstructure` functions, and introduces a Code of
Conduct.

The code of conduct that I chose to adopt is the Contributor Covenant.
This is an already existing code of conduct that is versioned and used
by other projects.