Add test for nesting values as well as support.
This started as a pretty small thing. I wanted to add some tests for
nesting values within one another. After I did this I realized that the
recursive part of this package just didn't work at all.

So in this commit I also fix those issues. It does so by checking the
type hints of the nested goal type rather than just checking the type of
the value.