Switch from Travis CI to Cirrus CI.
Travis was acquired and seems to sorta be on the shelf unfortunately. I
have more experience with Circle CI, but wanted to try something new. I
had tried Cirrus CI: https://cirrus-ci.org/ before with a Rust project
and was really pleasantly surprised.

This commit removes the Travis CI config and adds a Cirrus CI config.