Update all files to conform to linting.
In the previous commit we adding linting that wasn't passing. (be aware
for git bisect)

In this commit we deal with all of the linting error. Some linting
errors were skipped where it made sense. Some rules were adjusted.

Two files that changed quite a bit was the `structure.py` and
`unstructure.py` files which had many violations. `unstructure.py`
complained about too many returns so it required a significant refactor.